LPS-74270 Move to match kernel model dir
