+++ modules/apps/headless-apio/person/person-apio-impl/src/main/java/com/liferay/person/apio/internal/architect/router/OrganizationPersonNestedCollectionRouter.java
@@ -19,7 +19,7 @@ import com.liferay.apio.architect.pagination.Pagination;
 import com.liferay.apio.architect.router.NestedCollectionRouter;
 import com.liferay.apio.architect.routes.NestedCollectionRoutes;
 import com.liferay.organization.apio.architect.identifier.OrganizationIdentifier;
-import com.liferay.person.apio.internal.wrapper.UserWrapper;
+import com.liferay.person.apio.internal.model.UserWrapper;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.service.UserService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
